Add and use SeqRecord factory (from_Bio_SeqRecord)
Allows to easily convert `Bio.SeqRecord`s to `pydna.seqrecord.SeqRecord`s.
This is useful, for example, in the `__add__` method that previously returned a `Bio.SeqRecord`.
The code shared with the constructor that e.g. converts `str` attributes `to _pretty_str` is moved to a new method "`_fix_attributes`".